Support decimal values and SDK user agents (#4)
## Summary\n\n- preserve fractional price, cost, total-cost, and credit
values with the native decimal-capable type for each language\n- send
{SDK Name}/{SDK Version} as the default User-Agent, with a client
configuration override\n- bump the SDK version to 0.2.0\n\n##
Verification\n\n- formatting and lint targets pass\n- build and package
checks pass\n- request-capture and fractional-value smoke checks pass

+++ src/adsefid/models/user.py
@@ -16,7 +16,7 @@ from ..enums import LineSelector, TemplateParameterType, TemplateState
 class AccountInfo:
     name: str
     company_name: str
-    credit_left: int
+    credit_left: float
     email: str
     phone: str
     account_status: str
@@ -26,7 +26,7 @@ class AccountInfo:
         return cls(
             name=data["name"],
             company_name=data["company_name"],
-            credit_left=data["credit_left"],
+            credit_left=float(data["credit_left"]),
             email=data["email"],
             phone=data["phone"],
             account_status=data["account_status"],
